LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/processing-model-2/integration-with-the-javascript-agent-formalism/canblock-sharedworker-expected.txt

CONSOLE MESSAGE: line 12: ReferenceError: Can't find variable: SharedWorker

FAIL [[CanBlock]] in a shared worker agent ReferenceError: Can't find variable: SharedWorker
